Updated Cart API to add or subtract part. Fixed Cart route to add total value

+++ controllers/api/cart.js
@@ -2,33 +2,66 @@ const router = require('express').Router();
 const { request, response } = require('express');
 const { Product, CartProduct, Cart, User } = require('../../models');
 
-// Create a new cart
 
 
 
 // add new item to cart 
-router.post('/addCart', async (req, res) => {
+router.post('/:part_id', async (req, res) => {
     try {
-        const cartData = await CartProduct.create(req.body);
-        res.status(200).json(cartData);
+        let dbCartData = await CartProduct.create({ cart_id: req.session.userCart, product_id: req.params.part_id });
+        dbCartData = dbCartData.get({ plain: true });
+
+        if (!dbCartData) {
+            dbCartData = await CartProduct.create({ cart_id: req.session.userCart, product_id: req.params.part_id });
+        } else {
+            dbCartData = await CartProduct.update(
+                { qty: dbCartData.qty++ },
+                {
+                    where: {
+                        cart_id: req,
+                        product_id: req.params.part_id
+                    }
+                })
+        }
+
+        res.status(200).json(dbCartData);
     } catch (err) {
         res.status(400).json(err);
     }
 });
 
-
 // delete from cart
-router.delete('/cart/:part_id', async (req, res) => {
+router.delete('/:part_id', async (req, res) => {
     try {
         var dbCartData = await CartProduct.findOne({
             where: {
-                cart_id: req.session.cartID, 
-                product_id: req.params.part_id }});
-        dbCartData = await CartProduct.destroy({
-            where: {
-                id: dbCartData.id
+                cart_id: req.session.userCart,
+                product_id: req.params.part_id
             }
         });
+
+        dbCartData = dbCartData.get({ plain: true });
+
+
+        if (dbCartData.qty > 1) {
+            dbCartData = await CartProduct.update(
+                { qty: cartData.qty-- },
+                {
+                    where: {
+                        cart_id: req,
+                        product_id: req.params.part_id
+                    }
+                })
+            cartData = await CartProduct.create({ cart_id: req.session.userCart, product_id: req.params.part_id });
+        } else if (dbCartData == 1){
+            dbCartData = await CartProduct.destroy({
+                where: {
+                    cart_id: req,
+                    product_id: req.params.part_id
+                }
+            });
+        }
+
         return res.json(dbCartData);
     } catch (err) {
         console.log(err);
